редактирование кода

+++ app/src/main/java/by/bsuir/ppo_timer/DBHelper.java
@@ -7,7 +7,7 @@ import android.database.sqlite.SQLiteOpenHelper;
 public class DBHelper extends SQLiteOpenHelper {
     public DBHelper(Context context) {
         // конструктор суперкласса
-        super(context, "myDB", null, 3);
+        super(context, "myDB", null, 4);
     }
 
     @Override
@@ -15,13 +15,14 @@ public class DBHelper extends SQLiteOpenHelper {
         db.execSQL("create table workout ("
                 + "id integer primary key autoincrement,"
                 + "Name text,"
-                + "TimeOfPreparation text,"
-                + "TimeOfWork text,"
-                + "TimeOfRest text,"
-                + "CountOfCycles text,"
-                + "CountOfSets text,"
-                + "TimeOfRestBetweenSet text,"
-                + "TimeOfFinalRest text" + ");");
+                + "TimeOfPreparation integer,"
+                + "TimeOfWork integer,"
+                + "TimeOfRest integer,"
+                + "CountOfCycles integer,"
+                + "CountOfSets integer,"
+                + "TimeOfRestBetweenSet integer,"
+                + "TimeOfFinalRest integer,"
+                + "color integer" + ");");
     }
 
 
@@ -59,6 +60,45 @@ public class DBHelper extends SQLiteOpenHelper {
                         + "color integer" + ");");
 
                 db.execSQL("insert into workout select id, Name, TimeOfPreparation, TimeOfWork,TimeOfRest,CountOfCycles,CountOfSets,TimeOfRestBetweenSet,TimeOfFinalRest,color from people_tmp;");
+                db.execSQL("drop table people_tmp;");
+
+                db.setTransactionSuccessful();
+            } finally {
+                db.endTransaction();
+            }
+        }
+        if (oldVersion == 3 && newVersion == 4) {
+            db.beginTransaction();
+            try {
+
+                db.execSQL("create temporary table people_tmp ("
+                        + "id integer primary key autoincrement,"
+                        + "Name text,"
+                        + "TimeOfPreparation text,"
+                        + "TimeOfWork text,"
+                        + "TimeOfRest text,"
+                        + "CountOfCycles text,"
+                        + "CountOfSets text,"
+                        + "TimeOfRestBetweenSet text,"
+                        + "TimeOfFinalRest text,"
+                        + "color integer" + ");");
+
+                db.execSQL("insert into people_tmp select id, Name, TimeOfPreparation, TimeOfWork,TimeOfRest,CountOfCycles,CountOfSets,TimeOfRestBetweenSet,TimeOfFinalRest, color from workout;");
+                db.execSQL("drop table workout;");
+
+                db.execSQL("create table workout ("
+                        + "id integer primary key autoincrement,"
+                        + "Name text,"
+                        + "TimeOfPreparation integer,"
+                        + "TimeOfWork integer,"
+                        + "TimeOfRest integer,"
+                        + "CountOfCycles integer,"
+                        + "CountOfSets integer,"
+                        + "TimeOfRestBetweenSet integer,"
+                        + "TimeOfFinalRest integer,"
+                        + "color integer" + ");");
+
+                db.execSQL("insert into workout select id, Name, cast( TimeOfPreparation as integer), cast( TimeOfWork as integer),cast( TimeOfRest as integer),cast( CountOfCycles as integer),cast( CountOfSets as integer),cast( TimeOfRestBetweenSet as integer),cast( TimeOfFinalRest as integer),color from people_tmp;");
                 db.execSQL("drop table people_tmp;");
 
                 db.setTransactionSuccessful();
